"Canonicalize" class RunTimePoolInfo.
This class manages an important resource (an mmap() memory) region.
Rather than treating the class as a bundle of state with methods to
explicitly manipulate that resource, give it the full C++ concrete
data type treatment: Constructors, destructor, and move assignment, with
all data private.  Now the lifetime of the managed resource matches the
lifetime of the class instance that manages the resource: The
resource is freed when the instance is destroyed or assigned to.

Test: nn/runtime/test

Bug: 69685100
Bug: 69929024
Change-Id: I2514ed437e92f4210d241883883a9bffe75f68a1